move extract_qb_data to its own class
